requirements.txt
asgiref==3.9.1
cffi==1.17.1
cryptography==45.0.5
Django==5.2.4
djangorestframework==3.16.0
mysqlclient==2.2.7
pycparser==2.22
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
